pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "audiobook-dl"
authors = [
    { name = "Joakim Holm", email = "mail@joakimholm.xyz" }
]
description = "CLI tool for downloading audiobooks from online sources"
readme = "README.md"
keywords = ["audiobooks", "cli", "downloader"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Sound/Audio",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
]
dependencies = [
    "requests",
    "lxml",
    "rich",
    "mutagen",
    "pillow",
    "cssselect",
    "m3u8",
    "pycryptodome",
    "importlib-resources"
]
dynamic = ["version"]

[project.urls]
"Homepage" = "https://github.com/jo1gi/audiobook-dl"
"Bugtracker" = "https://github.com/jo1gi/audiobook-dl/issues"

[project.scripts]
audiobook-dl = "audiobookdl.__main__:run"

[tool.setuptools.dynamic]
version = {attr = "audiobookdl.__version__"}

[tool.setuptools.package-data]
mypkg = ["*.txt"]

[tool.mypy]
ignore_missing_imports = true
allow_untyped_globals = false
disallow_untyped_calls = true

[[tool.mypy.overrides]]
module = [
    "audiobookdl.sources.*",
    "audiobookdl.output.metadata.id3",
    "audiobookdl.output.metadata.mp4"
]
disallow_untyped_calls = false
